+++ sql/core/src/test/scala/org/apache/spark/sql/SparkSessionBuilderSuite.scala
@@ -152,4 +152,25 @@ class SparkSessionBuilderSuite extends SparkFunSuite with BeforeAndAfterEach {
       session.sparkContext.hadoopConfiguration.unset(mySpecialKey)
     }
   }
+
+  test("test SparkContext stopped when last SparkSession is stopped ") {
+    val session1 = SparkSession.builder()
+      .master("local")
+      .config(UI_ENABLED.key, value = false)
+      .config("some-config", "a")
+      .getOrCreate()
+
+    assert(!session1.sparkContext.isStopped)
+
+    val session2 = SparkSession.builder()
+      .master("local")
+      .config(UI_ENABLED.key, value = false)
+      .config("some-config", "b")
+      .getOrCreate()
+
+    session1.stop()
+    assert(!session1.sparkContext.isStopped)
+    session2.stop()
+    assert(session1.sparkContext.isStopped)
+  }
 }
